Revert "update ECS fargate collector to use v4 endpoint (#23253)"
This reverts commit 808c008ca500c582b31e0da20d72cc7db8c0439a.
