--- comp/core/workloadmeta/collectors/internal/ecsfargate/ecsfargate_test.go
@@ -1,21 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build docker
-
-package ecsfargate
-
-import (
-	"github.com/DataDog/datadog-agent/comp/core/workloadmeta"
-)
-
-type fakeWorkloadmetaStore struct {
-	workloadmeta.Component
-	notifiedEvents []workloadmeta.CollectorEvent
-}
-
-func (store *fakeWorkloadmetaStore) Notify(events []workloadmeta.CollectorEvent) {
-	store.notifiedEvents = append(store.notifiedEvents, events...)
-}
