--- comp/core/workloadmeta/collectors/internal/ecsfargate/testdata/redis.json
@@ -1,146 +1,0 @@
-{
-  "Cluster": "arn:aws:ecs:us-east-1:123457279990:cluster/ecs-cluster",
-  "TaskARN": "arn:aws:ecs:us-east-1:123457279990:task/ecs-cluster/938f6d263c464aa5985dc67ab7f38a7e",
-  "Family": "my-redis",
-  "Revision": "1",
-  "DesiredStatus": "RUNNING",
-  "KnownStatus": "RUNNING",
-  "Limits": { "CPU": 1, "Memory": 2048 },
-  "PullStartedAt": "2023-11-20T12:09:45.059013479Z",
-  "PullStoppedAt": "2023-11-20T12:10:41.166377771Z",
-  "AvailabilityZone": "us-east-1d",
-  "LaunchType": "FARGATE",
-  "Containers": [
-    {
-      "DockerId": "938f6d263c464aa5985dc67ab7f38a7e-1714341083",
-      "Name": "log_router",
-      "DockerName": "log_router",
-      "Image": "amazon/aws-for-fluent-bit:latest",
-      "ImageID": "sha256:ed2bd1c0fa887e59338a8761e040acc495213fd3c1b2be661c44c7158425e6e3",
-      "Labels": {
-        "com.amazonaws.ecs.container-name": "log_router",
-        "com.amazonaws.ecs.task-definition-family": "my-redis",
-        "com.amazonaws.ecs.task-definition-version": "1",
-        "com.amazonaws.ecs.cluster": "ecs-cluster"
-      },
-      "DesiredStatus": "RUNNING",
-      "KnownStatus": "RUNNING",
-      "Limits": { "CPU": 2 },
-      "CreatedAt": "2023-11-20T12:10:44.559880428Z",
-      "StartedAt": "2023-11-20T12:10:44.559880428Z",
-      "Type": "NORMAL",
-      "LogDriver": "awslogs",
-      "LogOptions": {
-        "awslogs-group": "aws",
-        "awslogs-region": "us-east-1",
-        "awslogs-stream": "log_router/log_router/938f6d263c464a"
-      },
-      "ContainerARN": "arn:aws:ecs:us-east-1:601427279990:container/ecs-cluster/938f6d263c464aa59/dc51359e-7f8a",
-      "Networks": [
-        {
-          "NetworkMode": "awsvpc",
-          "IPv4Addresses": ["172.31.15.128"],
-          "AttachmentIndex": 0,
-          "MACAddress": "0e:73:3c:72:d3:c6",
-          "IPv4SubnetCIDRBlock": "172.31.15.0/24",
-          "DomainNameServers": ["172.31.44.19", "172.31.0.2"],
-          "PrivateDNSName": "ip-123-31-115-123.ec2.internal",
-          "SubnetGatewayIpv4Address": "172.31.15.1/24"
-        }
-      ],
-      "Snapshotter": "overlayfs"
-    },
-    {
-      "DockerId": "938f6d263c464aa5985dc67ab7f38a7e-2537586469",
-      "Name": "datadog-agent",
-      "DockerName": "datadog-agent",
-      "Image": "public.ecr.aws/datadog/agent:latest",
-      "ImageID": "sha256:ba1d175ac08f8241d62c07785cbc6e026310cd2293dc4cf148e05d63655d1297",
-      "Labels": {
-        "com.amazonaws.ecs.container-name": "datadog-agent",
-        "com.amazonaws.ecs.task-definition-family": "my-redis",
-        "com.amazonaws.ecs.task-definition-version": "1"
-      },
-      "DesiredStatus": "RUNNING",
-      "KnownStatus": "RUNNING",
-      "Limits": { "CPU": 2 },
-      "CreatedAt": "2023-11-20T12:10:44.404563253Z",
-      "StartedAt": "2023-11-20T12:10:44.404563253Z",
-      "Type": "NORMAL",
-      "Health": {
-        "status": "HEALTHY",
-        "statusSince": "2023-11-20T12:11:16.383262018Z"
-      },
-      "Volumes": [
-        {
-          "DockerName": "my-redis-1-dd-sockets",
-          "Destination": "/var/run/datadog"
-        }
-      ],
-      "LogDriver": "awslogs",
-      "LogOptions": {
-        "awslogs-group": "aws",
-        "awslogs-region": "us-east-1",
-        "awslogs-stream": "log_router/datadog-agent/938f6d263c46e"
-      },
-      "ContainerARN": "arn:aws:ecs:us-east-1:601427279990:container/ecs-cluster/938f6d263c464aa/a17c293b-ab52",
-      "Networks": [
-        {
-          "NetworkMode": "awsvpc",
-          "IPv4Addresses": ["172.31.115.123"],
-          "AttachmentIndex": 0,
-          "MACAddress": "0e:73:3c:72:d3:ca",
-          "IPv4SubnetCIDRBlock": "172.31.11.0/24",
-          "DomainNameServers": ["172.31.44.17", "172.31.0.2"],
-          "PrivateDNSName": "ip-123-31-115-125.ec2.internal",
-          "SubnetGatewayIpv4Address": "172.31.11.1/24"
-        }
-      ],
-      "Snapshotter": "overlayfs"
-    },
-    {
-      "DockerId": "938f6d263c464aa5985dc67ab7f38a7e-3054012820",
-      "Name": "redis",
-      "DockerName": "redis",
-      "Image": "redis/redis:latest",
-      "ImageID": "sha256:529d1db9e246240208eab2a61fbb8614b09d7505ac1329c1dec70b6aba2e1428",
-      "Labels": {
-        "com.amazonaws.ecs.container-name": "redis",
-        "com.amazonaws.ecs.task-definition-family": "my-redis",
-        "com.amazonaws.ecs.task-definition-version": "1"
-      },
-      "DesiredStatus": "RUNNING",
-      "KnownStatus": "RUNNING",
-      "Limits": { "CPU": 2 },
-      "CreatedAt": "2023-11-20T12:11:16.701115523Z",
-      "StartedAt": "2023-11-20T12:11:16.701115523Z",
-      "Type": "NORMAL",
-      "LogDriver": "awslogs",
-      "LogOptions": {
-        "awslogs-group": "aws",
-        "awslogs-region": "us-east-1",
-        "awslogs-stream": "log_router/redis/938f6d263c4"
-      },
-      "ContainerARN": "arn:aws:ecs:us-east-1:601427279990:container/ecs-cluster/938f6d263c464aa5/a40518e2-7573",
-      "Networks": [
-        {
-          "NetworkMode": "awsvpc",
-          "IPv4Addresses": ["172.31.115.18"],
-          "AttachmentIndex": 0,
-          "MACAddress": "0e:73:3d:72:c3:cb",
-          "IPv4SubnetCIDRBlock": "172.31.15.0/24",
-          "DomainNameServers": ["172.31.44.19", "172.31.0.2"],
-          "PrivateDNSName": "ip-12-31-115-18.ec2.internal",
-          "SubnetGatewayIpv4Address": "172.111.115.1/24"
-        }
-      ],
-      "Snapshotter": "overlayfs"
-    }
-  ],
-  "ClockDrift": {
-    "ClockErrorBound": 0.5730080000000001,
-    "ReferenceTimestamp": "2023-12-27T16:06:41Z",
-    "ClockSynchronizationStatus": "SYNCHRONIZED"
-  },
-  "EphemeralStorageMetrics": { "Utilized": 2298, "Reserved": 20496 }
-}
